delivery-service/shipping-label-generate/requirements.txt
google-cloud-bigquery
functions-framework
flask==2.3.3